fix(factory): take variant as uint8 to match Rust's InvalidVariant() selector (BOP-168)
Change `createB20` and `getB20Address` to take `uint8 variant` (instead of
the `B20Variant` enum) so the factory body — not the ABI decoder — rejects
out-of-range bytes. The entry-point guard reverts with the typed
`InvalidVariant()` selector, matching the Rust precompile's
`B20Variant::from_abi(call.variant).ok_or_else(... InvalidVariant {})`
behavior on both create and address-prediction paths.

The wire-level 4-byte selectors are unchanged (Solidity normalizes enum
parameters to their underlying `uint8` type at the canonical signature
level), so no client outside the Solidity source tree is affected. In-tree
Solidity callers add explicit `uint8(IB20Factory.B20Variant.X)` casts at
call sites.

The dead `else { revert InvalidVariant(); }` branch in `createB20`'s
dispatch chain now serves as defense-in-depth: any future enum variant
added without a corresponding dispatch arm will hit this branch instead
of being silently routed to ASSET.

Tightens the two parity tests to pin `vm.expectRevert(IB20Factory.InvalidVariant.selector)`
instead of bare `vm.expectRevert()`, so a future regression to the
decode-time `Panic(0x21)` would fail the test.

+++ test/lib/B20FactoryTest.sol
@@ -111,7 +111,7 @@ contract B20FactoryTest is BaseTest {
         bytes[] memory initCalls
     ) internal returns (address token) {
         vm.prank(caller);
-        return factory.createB20(IB20Factory.B20Variant.DEFAULT, salt, abi.encode(params), initCalls);
+        return factory.createB20(uint8(IB20Factory.B20Variant.DEFAULT), salt, abi.encode(params), initCalls);
     }
 
     /// @notice Create a default-variant token with defaults (alice creator, fresh salt, empty init calls).
@@ -127,7 +127,7 @@ contract B20FactoryTest is BaseTest {
         bytes[] memory initCalls
     ) internal returns (address token) {
         vm.prank(caller);
-        return factory.createB20(IB20Factory.B20Variant.STABLECOIN, salt, abi.encode(params), initCalls);
+        return factory.createB20(uint8(IB20Factory.B20Variant.STABLECOIN), salt, abi.encode(params), initCalls);
     }
 
     /// @notice Create a stablecoin-variant token with defaults.
@@ -143,7 +143,7 @@ contract B20FactoryTest is BaseTest {
         bytes[] memory initCalls
     ) internal returns (address token) {
         vm.prank(caller);
-        return factory.createB20(IB20Factory.B20Variant.ASSET, salt, abi.encode(params), initCalls);
+        return factory.createB20(uint8(IB20Factory.B20Variant.ASSET), salt, abi.encode(params), initCalls);
     }
 
     /// @notice Create a security-variant token with defaults.
